Added primitives

+++ FunksjonellProgrammering.Api/CreateUser/Extensions.cs
@@ -8,5 +8,5 @@ public static class Extensions
         => new (new Name(d.Name), d.Role);
 
     public static Entity ToEntity(this Domain d)
-        => new(d.Name.Value, d.Role);
+        => new(d.Name.Value, (int)d.Role);
 }
